updated old dialogs
