+++ package.json
@@ -74,7 +74,6 @@
         "@types/react": "^18.0.0",
         "big.js": "^6.2.1",
         "clsx": "^1.2.1",
-        "is-url": "^1.2.4",
         "js-base64": "^3.7.2",
         "lodash.debounce": "^4.0.8",
         "near-api-js": "^0.45.1",
@@ -85,8 +84,7 @@
         "react-dom": "^18.0.0",
         "react-router-dom": "6",
         "regenerator-runtime": "^0.13.9",
-        "rxjs": "^7.5.6",
-        "validator": "^13.7.0"
+        "rxjs": "^7.5.6"
     },
     "resolutions": {
         "near-api-js": "0.45.1"
